requirements.txt
Flask
Flask-CORS
mysql-connector-python
pymongo
PyJWT
fastapi 
uvicorn 
python-dotenv 
bcrypt 
jwt 
fastapi-mail 
itsdangerous
python-jose